backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.1
redis>=5.0.4
pydantic[email]>=2.7.1
pydantic-settings>=2.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
httpx>=0.27.0
torch>=2.3.0
torchvision>=0.18.0
transformers>=4.41.0
librosa>=0.10.2
opencv-python-headless>=4.9.0.80
numpy>=1.26.4
scipy>=1.13.0
scikit-learn>=1.5.0
Pillow>=10.3.0
aiofiles>=23.2.1
python-dotenv>=1.0.1
authlib>=1.3.0
aiosmtplib>=3.0.1
imbalanced-learn>=0.12.2
bcrypt==4.0.1
certifi
groq>=0.9.0
langdetect>=1.0.9
deep-translator>=1.11.4
